pyproject.toml
[project]
name = "simple-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30",
    "loguru>=0.7.3",
    "mypy>=1.19.1",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
    "ruff>=0.15.8",
    "sqlalchemy[asyncio]>=2.0",
    "types-pyyaml>=6.0.12.20250915",
]
